Assignment_Two/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
lxml==5.3.0
outcome==1.3.0.post0
pydantic==2.10.6
pydantic_core==2.27.2
pyperclip==1.9.0
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
responses==0.24.1
selenium==4.28.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
